requirements_py3.txt
# python -m pip install --upgrade pip
# pip install torch torchvision torchaudio torchmetrics --index-url https://download.pytorch.org/whl/cu126

ipykernel
ipython
ipywidgets
matplotlib
pandas
pillow
requests
scikit-learn
scipy
seaborn
kagglehub
soundfile
detecto
ultralytics
torch_geometric
hugsvision
tensorboard
accelerate>=0.26.0
datasets